update: 更新文件 i_bluetooth_ble_advertiser.h

+++ frameworks/inner/ipc/interface/i_bluetooth_ble_advertiser.h
@@ -44,7 +44,7 @@ public:
         const BluetoothBleAdvertiserData &scanResponse, int32_t advHandle,
         bluetooth::SetAdvDataType type = bluetooth::SET_ADV_DATA_BOTH) = 0;
     virtual int ChangeAdvertisingParams(uint8_t advHandle, const BluetoothBleAdvertiserSettings &settings) = 0;
-    virtual int8_t BleStartRangeAdv(bluetooth::BleAppType appType) = 0;
+    virtual int BleStartRangeAdv(bluetooth::BleAppType appType) = 0;
     virtual int BleStopRangeAdv(bluetooth::BleAppType appType) = 0;
 };
 }  // namespace Bluetooth
